+++ scripts/solve_network.py
@@ -1571,6 +1571,9 @@ def collect_kwargs(
     model_kwargs["linearized_unit_commitment"] = cf_solving.get(
         "linearized_unit_commitment", False
     )
+    model_kwargs["include_objective_constant"] = cf_solving.get(
+        "include_objective_constant"
+    )
 
     # Solve kwargs
     solver_name = solving["solver"]["name"]